# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": "0.1"
    },
    "paths": {
...
        "/animes": {
            "get": {
                "tags": [
                    "Animes"
                ],
                "summary": "Retorna uma lista de animes com ou sem paginação.",
                "description": "Este endpoint pode retornar todos os animes ou uma lista paginada, dependendo do parâmetro 'por_pagina'. Exemplo de requisição curl: curl -X GET http://127.0.0.1:8000/api/animes -H Authorization:Bearer SEU_TOKEN",
                "operationId": "6b9b19c24358ee0b2a42e45472e7c6b2",
                "parameters": [
                    {
                        "name": "por_pagina",
                        "in": "query",
                        "description": "Número de itens por página para a paginação.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de animes retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titulo": {
                                                "type": "string",
                                                "example": "Naruto"
                                            },
                                            "genero": {
                                                "type": "string",
                                                "example": "Aventura, Ação"
                                            },
                                            "episodios": {
                                                "type": "integer",
                                                "example": 500
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor."
                    }
                }
            },
            "post": {
                "tags": [
                    "Animes"
                ],
                "summary": "Cria um novo anime.",
                "description": "Cria um novo registro de anime no banco de dados com base nos dados fornecidos.",
                "operationId": "cae91423263c61ef819fa39116a7430e",
                "requestBody": {
                    "description": "Dados do anime a ser criado.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "sinopse",
                                    "ano_lancamento",
                                    "id_status",
                                    "url_imagem"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Demon Slayer"
                                    },
                                    "sinopse": {
                                        "description": "Sinopse do anime",
                                        "type": "string",
                                        "example": "Tanjiro Kamado é um jovem bondoso que se torna um caçador de demônios após sua família ser massacrada por demônios e sua irmã Nezuko ser transformada em um deles."
                                    },
                                    "ano_lancamento": {
                                        "description": "Ano de lançamento do anime",
                                        "type": "integer",
                                        "example": 2019
                                    },
                                    "id_status": {
                                        "description": "ID do status do anime",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "url_imagem": {
                                        "description": "Imagem do anime (jpeg, png, jpg, gif ou webp, máx 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anime criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "Demon Slayer"
                                        },
                                        "sinopse": {
                                            "type": "string",
                                            "example": "Tanjiro Kamado é um jovem bondoso que se torna um caçador de demônios após sua família ser massacrada por demônios e sua irmã Nezuko ser transformada em um deles."
                                        },
                                        "ano_lancamento": {
                                            "type": "integer",
                                            "example": 2019
                                        },
                                        "id_status": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "O campo titulo é obrigatório."
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
...
        "/animes/{animeId}/vincular-categoria": {
            "post": {
                "tags": [
                    "Animes"
                ],
                "summary": "Vincula uma categoria a um anime.",
                "description": "Associa uma categoria existente a um anime específico. O ID do anime é passado na URL, e o ID da categoria é enviado no corpo da requisição.",
                "operationId": "2379c09e4be9d9c11990530a656af7c2",
                "parameters": [
                    {
                        "name": "animeId",
                        "in": "path",
                        "description": "ID do anime ao qual a categoria será vinculada.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "ID da categoria a ser vinculada.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categoria_id"
                                ],
                                "properties": {
                                    "categoria_id": {
                                        "description": "ID da categoria.",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria vinculada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria vinculada com sucesso."
                                        },
                                        "anime_categoria": {
                                            "properties": {
                                                "anime_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "categoria_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "categoria": {
                                                    "description": "Detalhes da categoria vinculada.",
                                                    "type": "object"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Anime ou Categoria não encontrados.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Anime ou categoria não encontrados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo categoria_id é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
...
    },
...
}